// redux/actions/types.ts
export enum actions {
  setDate = 'SET_DATE',
  setType = 'SET_TYPE',
  setCategory = 'SET_CATEGORY',
  setHorse = 'SET_HORSE',
  setAdmin = 'SET_ADMIN',
  setUser = 'SET_USER',
  setTime = 'SET_TIME',
  setIsOpen = 'SET_IS_OPEN',
  setBookingIsOpen = 'SET_BOOKING_IS_OPEN',
  resetStore = 'RESET_STORE',
  setCount = 'SET_COUNT',
}